requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.7
django-cors-headers==4.7.0
django-unfold==0.53.0
djangorestframework==3.15.2
msgpack==1.1.0
redis==5.2.1
sqlparse==0.5.3
tzdata==2025.2
